+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Added
 - Add foreign field renderer that renders object with `a` tag
 - Add Json field renderer
 - Add action column to list view with actions view,edit and delete (remove row click)
+- Add graph dashboard widget + improve widget options
 
 Changed
 ~~~~~~~